// prisma/generated/schemas/objects/NestedUuidNullableWithAggregatesFilter.schema.ts
import type { Prisma } from "@prisma/client"
import { z } from "zod"

import { NestedIntNullableFilterObjectSchema } from "./NestedIntNullableFilter.schema"
import { NestedStringNullableFilterObjectSchema } from "./NestedStringNullableFilter.schema"

const Schema: z.ZodType<Prisma.NestedUuidNullableWithAggregatesFilter> = z
  .object({
    equals: z.string().optional().nullable(),
    in: z.union([z.string().array(), z.string()]).optional().nullable(),
    notIn: z.union([z.string().array(), z.string()]).optional().nullable(),
    lt: z.string().optional(),
    lte: z.string().optional(),
    gt: z.string().optional(),
    gte: z.string().optional(),
    not: z
      .union([
        z.string(),
        z.lazy(() => NestedUuidNullableWithAggregatesFilterObjectSchema),
      ])
      .optional()
      .nullable(),
    _count: z.lazy(() => NestedIntNullableFilterObjectSchema).optional(),
    _min: z.lazy(() => NestedStringNullableFilterObjectSchema).optional(),
    _max: z.lazy(() => NestedStringNullableFilterObjectSchema).optional(),
  })
  .strict()

export const NestedUuidNullableWithAggregatesFilterObjectSchema = Schema